Moving get_web3_and_hyperdrive_contracts from agent0 to ethpy
