+++ lib/agent0/agent0/hyperdrive/exec/__init__.py
@@ -12,5 +12,5 @@ from .execute_agent_trades import (
 from .fund_agents import fund_agents
 from .get_agent_accounts import get_agent_accounts
 from .run_agents import run_agents
-from .setup_experiment import get_web3_and_contracts, register_username, setup_experiment
+from .setup_experiment import register_username, setup_experiment
 from .trade_loop import get_wait_for_new_block, trade_if_new_block
