+++ lib/ethpy/ethpy/hyperdrive/__init__.py
@@ -2,6 +2,7 @@
 from .addresses import HyperdriveAddresses, fetch_hyperdrive_address_from_url
 from .assets import AssetIdPrefix, decode_asset_id, encode_asset_id
 from .errors import HyperdriveErrors, lookup_hyperdrive_error_selector
+from .get_web3_and_hyperdrive_contracts import get_web3_and_hyperdrive_contracts
 from .interface import (
     get_hyperdrive_checkpoint_info,
     get_hyperdrive_config,
